backend/requirements.txt
fastapi>=0.119.0
uvicorn[standard]>=0.24.0
pydantic>=2.7.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
python-dotenv==1.0.0
openai>=1.6.1
langchain==0.0.350
langchain-openai==0.0.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
httpx==0.25.2
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
